+++ tools/check.sh
@@ -17,3 +17,4 @@ mkdir -p java/build
 javac --release 8 -d java/build $(find java/src -name '*.java')
 javac --release 8 -cp java/build -d java/build $(find java/test -name '*.java')
 java -cp java/build fr.iamacat.spi.SkeletonCheck
+java -cp java/build fr.iamacat.bridge.BridgeCheck